requirements.txt
langextract
langchain-core
langchain-community
python-dotenv
